Refactor RK4 integration
The nature of RK4 integration requires a bit of a rethink into how the particle updates are parallelized. I've refactored both the Euler and RK4 updateVelocity functions to place the parallelization inside the scope of the functions themselves. Assuming I've done this correctly, this should now be able to correctly implement RK4, as well as providing additional flexibility if/when you decide to try more integrators.